+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:39701/7ce058fd-86c4-49d1-b1a9-835077fe4d8a?"}
+{"mongoUri":"mongodb://127.0.0.1:43431/fd217de3-0918-46dc-b79a-7e9d5d3e4419?"}
